fix(batchnorm): impl  static_tensors

+++ src/nn/embedding.rs
@@ -10,7 +10,7 @@ use super::{Module, StateDict, Trainable};
 /// This layer is used to convert a sequence of integers into a sequence of one-hot vectors.
 #[derive(Debug, CallableModule, NonParameterModule)]
 pub struct OneHot {
-    pub num_classes: u32,
+    pub num_classes: i64,
 }
 
 impl Module for OneHot {
@@ -24,7 +24,7 @@ impl Module for OneHot {
 }
 
 impl OneHot {
-    pub fn new(num_classes: u32) -> Self {
+    pub fn new(num_classes: i64) -> Self {
         Self { num_classes }
     }
 }
@@ -56,7 +56,7 @@ impl Embedding {
             num_embeddings,
             embedding_dim,
             weight: weight.cell(),
-            one_hot: OneHot::new(num_embeddings as u32),
+            one_hot: OneHot::new(num_embeddings as i64),
         }
     }
 }
